Fixed Runtime.js to not always remove last argument
Fixed a bug where if you used any Async method in a standard await function, it would always remove the last argument of the function you are trying to call (if you supplied any)
Which isn't supposed to happen.

This doesn't happen if you use the default Async method, via the water-fall method, where the function you are trying to call is embedded to the call itself.

Now supports both methods!
1. Water-fall (call-back embedded)
2. Standard await.